requirements.txt
adapt-parser
certifi==2017.4.17
chardet==3.0.4
configobj==5.0.6
idna==2.5
mustache==0.1.4
pyee==1.0.1
requests==2.18.1
six==1.10.0
todoist-python==7.0.17
urllib3==1.21.1
websocket-client==0.32.0
